Have coded URI in PHP that returns crime.csv file in RDF formatted code.
Have used JSON-LD as as the RDF format to return the machine readable code that represents the data from the crime.csv file. This code that produces the RDF format is within the JSON-LD.php file. Will link this .php page to the machine readable (JSON-LD) button on the index.php page. Will code comment elements on this page soon.